+++ R/mosaic_tiles.R
@@ -20,16 +20,22 @@ mosaic_tiles <- function(pattern, dir_path, vrt_save, siteName = NULL,
   output_vrt_path <- file.path(getwd(), paste0(siteName, pattern,'_mosaic.vrt'))
   if (!file.exists(output_vrt_path))
     v <- terra::vrt(x = listfiles, filename = output_vrt_path)
+
   # create tiff from vrt
   mosaic_path <- file.path(dir_path, paste0(siteName, pattern,'_mosaic.tiff'))
   if (!file.exists(mosaic_path) | overwrite)
     sf::gdal_utils(util = 'translate', source = output_vrt_path,
                    destination = mosaic_path,
                    options = c("-co", "COMPRESS=LZW"))
-  # create tiff from vrt
-  dir.create(path = vrt_save, showWarnings = FALSE, recursive = TRUE)
-  output_vrt_path2 <- file.path(vrt_save,
-                                paste0(siteName, pattern,'_mosaic.vrt'))
-  file.rename(from = output_vrt_path, to = output_vrt_path2)
+
+  # delete vrt
+  file.remove(output_vrt_path)
+  # dir.create(path = vrt_save, showWarnings = FALSE, recursive = TRUE)
+  # output_vrt_path2 <- file.path(vrt_save,
+  #                               paste0(siteName, pattern,'_mosaic.vrt'))
+  # file.rename(from = output_vrt_path, to = output_vrt_path2)
+
+  # delete individual files
+  file.remove(listfiles)
   return(mosaic_path)
 }